pyproject.toml
[project]
name = "2getherapp"
version = "0.1.0"
description = ""
authors = [
    {name = "EldarJc",email = "jusiceldar87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
